setup: Use the NetInfo address if given
If the user specifies the afs_server_netinfo to set the NetInfo file,
then use that address in the server's CellServDB as well.  The NetInfo
can have multiple addresses, but for this test setup, just support a
single address to keep this simple.

For example, you can create a dummy-interface with a persisent address
and use that address for the test cell with the afs_server_netinfo
variable.